+++ hodor-core/src/main/java/org/dromara/hodor/core/HodorMetadata.java
@@ -13,7 +13,7 @@ import lombok.Data;
 public class HodorMetadata {
 
     private List<String> nodes;
-    private List<Integer> interval;
+    private List<Long> interval;
     private List<CopySet> copySets;
 
 }
